+++ packages/terra-core-docs/src/terra-dev-site/doc/button/example/ButtonIcon.jsx
@@ -6,13 +6,15 @@ import styles from './ButtonDocCommon.module.scss';
 
 const cx = classNames.bind(styles);
 
-const Icon = <IconEdit />;
+const InformativeIcon = <IconEdit a11yLabel="Edit Icon" />;
+const DecorativeIcon = <IconEdit />;
 
 const ButtonIcon = () => (
   <div>
-    <Button icon={Icon} text="icon" className={cx('button')} />
-    <Button icon={Icon} text="icon" isReversed className={cx('button')} />
-    <Button icon={Icon} isIconOnly text="Icon Only Button" className={cx('button')} />
+    <Button icon={InformativeIcon} text="Informative icon" className={cx('button')} />
+    <Button icon={DecorativeIcon} text="Decorative icon" isReversed className={cx('button')} />
+    <Button icon={DecorativeIcon} isIconOnly text="Decorative Icon Only Button" className={cx('button')} />
+    <Button icon={InformativeIcon} isIconOnly text="Informative Icon Only Button" className={cx('button')} />
   </div>
 );